requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
gdown